Added new files for lists, dicts, events and dungeon related things for easier accessibility and readability. Furthermore, the README.md file was updated to reflect the new structure of the project.
