+++ events.py
@@ -1,0 +1,27 @@
+
+import random
+
+from classes import *
+from lists_and_dicts import *
+from dungeon import *
+
+events = [
+    "event_enemy_encounter",
+    "event_wandering_trader_encounter",
+    "event_player_finds_village",
+    "event_player_finds_dungeon",
+    "event_player_finds_random_house",
+    "event_npc_encounter"
+    "event_player_finds_ruins",
+    "event_payer_finds_ancient_ruins",
+    "event_player_finds_camp",
+    "event_player_finds_hostile_camp",
+    "event_player_random_thoughts",
+    "event_bird_shits_on_player",
+    "event_random_animal",
+    "event_player_finds_farmhouse",
+    "event_player_finds_abandoned_farmhouse",
+    "event_player_finds_herb",
+    "event_player_finds_city",
+    "event_player_finds_npc_getting_attacked",
+]
